Indirect gh functions for octo/commands.lua
