+++ lua/octo/backend/gh/commands.lua
@@ -1,13 +1,17 @@
 local utils = require "octo.utils"
 local cli = require "octo.backend.gh.cli"
 local graphql = require "octo.backend.gh.graphql"
--- local constants = require "octo.constants"
+local constants = require "octo.constants"
 local window = require "octo.ui.window"
 local writers = require "octo.ui.writers"
 
 local _, Job = pcall(require, "plenary.job")
 
 local M = {}
+
+-- -----------------------------------------------------------------------------
+-- Functions for init.lua
+-- -----------------------------------------------------------------------------
 
 ---@param repo string
 ---@param kind string
@@ -108,7 +112,7 @@ function M.user_popup(login)
 end
 
 ---@param repo string
----@param number integer
+---@param number integer IID
 function M.link_popup(repo, number)
   local owner, name = utils.split_repo(repo)
   local query = graphql("issue_summary_query", owner, name, number)
@@ -135,12 +139,731 @@ function M.link_popup(repo, number)
 end
 
 
+-- -----------------------------------------------------------------------------
+-- Functions for commands.lua
+-- -----------------------------------------------------------------------------
 
+---@param comment CommentMetadata
+---@param buffer OctoBuffer
+---@param bufnr integer
+function M.cmds_delete_comment(comment, buffer, bufnr)
+  local start_line = comment.bufferStartLine
+  local end_line = comment.bufferEndLine
+  local query, threadId
+  if comment.kind == "IssueComment" then
+    query = graphql("delete_issue_comment_mutation", comment.id)
+  elseif comment.kind == "PullRequestReviewComment" then
+    query = graphql("delete_pull_request_review_comment_mutation", comment.id)
+    local _thread = buffer:get_thread_at_cursor()
+    threadId = _thread.threadId
+  elseif comment.kind == "PullRequestReview" then
+    -- Review top level comments cannot be deleted here
+    return
+  end
+  local choice = vim.fn.confirm("Delete comment?", "&Yes\n&No\n&Cancel", 2)
+  if choice == 1 then
+    cli.run {
+      args = { "api", "graphql", "-f", string.format("query=%s", query) },
+      cb = function(output)
+        -- TODO: deleting the last review thread comment, it deletes the whole thread and review
+        -- In issue buffers, we should hide the thread snippet
+        local resp = vim.fn.json_decode(output)
 
+        -- remove comment lines from the buffer
+        if comment.reactionLine then
+          vim.api.nvim_buf_set_lines(bufnr, start_line - 2, end_line + 1, false, {})
+          vim.api.nvim_buf_clear_namespace(bufnr, constants.OCTO_REACTIONS_VT_NS, start_line - 2, end_line + 1)
+        else
+          vim.api.nvim_buf_set_lines(bufnr, start_line - 2, end_line - 1, false, {})
+        end
+        vim.api.nvim_buf_clear_namespace(bufnr, comment.namespace, 0, -1)
+        vim.api.nvim_buf_del_extmark(bufnr, constants.OCTO_COMMENT_NS, comment.extmark)
+        local comments = buffer.commentsMetadata
+        if comments then
+          local updated = {}
+          for _, c in ipairs(comments) do
+            if c.id ~= comment.id then
+              table.insert(updated, c)
+            end
+          end
+          buffer.commentsMetadata = updated
+        end
 
+        if comment.kind == "PullRequestReviewComment" then
+          local review = require("octo.reviews").get_current_review()
+          if not review then
+            utils.error "Cannot find review for this comment"
+            return
+          end
 
+          local threads = resp.data.deletePullRequestReviewComment.pullRequestReview.pullRequest.reviewThreads.nodes
 
+          -- check if there is still at least a PENDING comment
+          local review_was_deleted = true
+          for _, thread in ipairs(threads) do
+            for _, c in ipairs(thread.comments.nodes) do
+              if c.state == "PENDING" then
+                review_was_deleted = false
+                break
+              end
+            end
+          end
+          if review_was_deleted then
+            -- we deleted the last pending comment and therefore GitHub closed the review, create a new one
+            review:create(function(resp)
+              review.id = resp.data.addPullRequestReview.pullRequestReview.id
+              local updated_threads = resp.data.addPullRequestReview.pullRequestReview.pullRequest.reviewThreads.nodes
+              review:update_threads(updated_threads)
+            end)
+          else
+            review:update_threads(threads)
+          end
 
+          -- check if we removed the last comment of a thread
+          local thread_was_deleted = true
+          for _, thread in ipairs(threads) do
+            if threadId == thread.id then
+              thread_was_deleted = false
+              break
+            end
+          end
+          if thread_was_deleted then
+            -- this was the last comment, close the thread buffer
+            -- No comments left
+            utils.error("Deleting buffer " .. tostring(bufnr))
+            local bufname = vim.api.nvim_buf_get_name(bufnr)
+            local split = string.match(bufname, "octo://.+/review/[^/]+/threads/([^/]+)/.*")
+            if split then
+              local layout = require("octo.reviews").get_current_review().layout
+              local file = layout:cur_file()
+              local diff_win = file:get_win(split)
+              local thread_win = file:get_alternative_win(split)
+              local original_buf = file:get_alternative_buf(split)
+              -- move focus to the split containing the diff buffer
+              -- restore the diff buffer so that window is not closed when deleting thread buffer
+              vim.api.nvim_win_set_buf(thread_win, original_buf)
+              -- delete the thread buffer
+              pcall(vim.api.nvim_buf_delete, bufnr, { force = true })
+              -- refresh signs and virtual text
+              file:place_signs()
+              -- diff buffers
+              file:show_diff()
+            end
+          end
+        end
+      end,
+    }
+  end
+end
 
+---@param bufnr integer
+---@param thread table Graphql Reponse Thread
+---@param thread_id string
+---@param thread_line integer
+local function update_review_thread_header(bufnr, thread, thread_id, thread_line)
+  local start_line = thread.originalStartLine ~= vim.NIL and thread.originalStartLine or thread.originalLine
+  local end_line = thread.originalLine
+  local commit_id = ""
+  for _, review_threads in ipairs(thread.pullRequest.reviewThreads.nodes) do
+    if review_threads.id == thread_id then
+      commit_id = review_threads.comments.nodes[1].originalCommit.abbreviatedOid
+    end
+  end
+  writers.write_review_thread_header(bufnr, {
+    path = thread.path,
+    start_line = start_line,
+    end_line = end_line,
+    commit = commit_id,
+    isOutdated = thread.isOutdated,
+    isResolved = thread.isResolved,
+  }, thread_line - 2)
+  local threads = thread.pullRequest.reviewThreads.nodes
+  local review = require("octo.reviews").get_current_review()
+  if review then
+    review:update_threads(threads)
+  end
+end
+
+---@param thread ThreadMetadata
+---@param bufnr integer
+function M.cmds_resolve_thread(thread, bufnr, _)
+  local thread_id = thread.threadId
+  local thread_line = thread.bufferStartLine
+
+  local query = graphql("resolve_review_thread_mutation", thread_id)
+  cli.run {
+    args = { "api", "graphql", "-f", string.format("query=%s", query) },
+    cb = function(output, stderr)
+      if stderr and not utils.is_blank(stderr) then
+        utils.error(stderr)
+      elseif output then
+        local resp = vim.fn.json_decode(output)
+        local resp_thread = resp.data.resolveReviewThread.thread
+        if resp_thread.isResolved then
+          update_review_thread_header(bufnr, resp_thread, thread_id, thread_line)
+          --vim.cmd(string.format("%d,%dfoldclose", thread_line, thread_line))
+        end
+      end
+    end,
+  }
+end
+
+---@param thread ThreadMetadata
+---@param bufnr integer
+function M.cmds_unresolve_thread(thread, bufnr, _)
+  local thread_id = thread.threadId
+  local thread_line = thread.bufferStartLine
+
+  local query = graphql("unresolve_review_thread_mutation", thread_id)
+  cli.run {
+    args = { "api", "graphql", "-f", string.format("query=%s", query) },
+    cb = function(output, stderr)
+      if stderr and not utils.is_blank(stderr) then
+        utils.error(stderr)
+      elseif output then
+        local resp = vim.fn.json_decode(output)
+        local resp_thread = resp.data.unresolveReviewThread.thread
+        if not resp_thread.isResolved then
+          update_review_thread_header(bufnr, resp_thread, thread_id, thread_line)
+        end
+      end
+    end,
+  }
+end
+
+---@param buffer OctoBuffer
+---@param bufnr integer
+---@param state string CLOSED or OPEN
+function M.cmds_change_state(buffer, bufnr, state)
+  local id = buffer.node.id
+  local query
+  if buffer:isIssue() then
+    query = graphql("update_issue_state_mutation", id, state)
+  elseif buffer:isPullRequest() then
+    query = graphql("update_pull_request_state_mutation", id, state)
+  end
+
+  cli.run {
+    args = { "api", "graphql", "-f", string.format("query=%s", query) },
+    cb = function(output, stderr)
+      if stderr and not utils.is_blank(stderr) then
+        utils.error(stderr)
+      elseif output then
+        local resp = vim.fn.json_decode(output)
+        local new_state, obj
+        if buffer:isIssue() then
+          obj = resp.data.updateIssue.issue
+          new_state = obj.state
+        elseif buffer:isPullRequest() then
+          obj = resp.data.updatePullRequest.pullRequest
+          new_state = obj.state
+        end
+        if state == new_state then
+          buffer.node.state = new_state
+          writers.write_state(bufnr, new_state:upper(), buffer.number)
+          writers.write_details(bufnr, obj, true)
+          utils.info("Issue state changed to: " .. new_state)
+        end
+      end
+    end,
+  }
+end
+
+---@param repo string
+---@param title string
+---@param body string
+function M.cmds_save_issue(repo, title, body)
+  local repo_id = utils.get_repo_id(repo)
+  local query = graphql("create_issue_mutation", repo_id, title, body)
+
+  cli.run {
+    args = { "api", "graphql", "-f", string.format("query=%s", query) },
+    cb = function(output, stderr)
+      if stderr and not utils.is_blank(stderr) then
+        utils.error(stderr)
+      elseif output then
+        local resp = vim.fn.json_decode(output)
+        require("octo").create_buffer("issue", resp.data.createIssue.issue, repo, true)
+        vim.fn.execute "normal! Gk"
+        vim.fn.execute "startinsert"
+      end
+    end,
+  }
+end
+
+---@param candidates table
+---@param repo string
+---@param repo_idx integer
+---@param title string
+---@param body string
+---@param base_ref_name string
+---@param head_ref_name string
+---@param is_draft boolean
+function M.cmds_save_pr(candidates, repo, repo_idx, title, body, base_ref_name, head_ref_name, is_draft)
+  local repo_id = utils.get_repo_id(candidates[repo_idx])
+  title = title and title or ""
+  body = body and body or ""
+  local query = graphql(
+    "create_pr_mutation",
+    base_ref_name,
+    head_ref_name,
+    repo_id,
+    utils.escape_char(title),
+    utils.escape_char(body),
+    is_draft
+  )
+
+  local choice = vim.fn.confirm("Create PR?", "&Yes\n&No\n&Cancel", 2)
+  if choice == 1 then
+    cli.run {
+      args = { "api", "graphql", "-f", string.format("query=%s", query) },
+      cb = function(output, stderr)
+        if stderr and not utils.is_blank(stderr) then
+          utils.error(stderr)
+        elseif output then
+          local resp = vim.fn.json_decode(output)
+          local pr = resp.data.createPullRequest.pullRequest
+          utils.info(string.format("#%d - `%s` created successfully", pr.number, pr.title))
+          require("octo").create_buffer("pull", pr, repo, true)
+          vim.fn.execute "normal! Gk"
+          vim.fn.execute "startinsert"
+        end
+      end,
+    }
+  end
+end
+
+---@param bufnr integer
+---@param number integer MR IID
+function M.cmds_mark_pr_ready(bufnr, number)
+  cli.run {
+    args = { "pr", "ready", tostring(number) },
+    cb = function(output, stderr)
+      utils.info(output)
+      utils.error(stderr)
+      writers.write_state(bufnr)
+    end,
+  }
+end
+
+---@param bufnr integer
+---@param number integer MR IID
+function M.cmds_mark_pr_draft(bufnr, number)
+  cli.run {
+    args = { "pr", "ready", tostring(number), "--undo" },
+    cb = function(output, stderr)
+      utils.info(output)
+      utils.error(stderr)
+      writers.write_state(bufnr)
+    end,
+  }
+end
+
+---@param buffer OctoBuffer
+function M.cmds_pr_checks(buffer)
+  cli.run {
+    args = { "pr", "checks", tostring(buffer.number), "-R", buffer.repo },
+    cb = function(output, stderr)
+      if stderr and not utils.is_blank(stderr) then
+        utils.error(stderr)
+      elseif output then
+        local max_lengths = {}
+        local parts = {}
+        for _, l in pairs(vim.split(output, "\n")) do
+          local line_parts = vim.split(l, "\t")
+          for i, p in pairs(line_parts) do
+            if max_lengths[i] == nil or max_lengths[i] < #p then
+              max_lengths[i] = #p
+            end
+          end
+          table.insert(parts, line_parts)
+        end
+        local lines = {}
+        for _, p in pairs(parts) do
+          local line = {}
+          for i, pp in pairs(p) do
+            table.insert(line, pp .. (" "):rep(max_lengths[i] - #pp))
+          end
+          table.insert(lines, table.concat(line, "  "))
+        end
+        local _, wbufnr = window.create_centered_float {
+          header = "Checks",
+          content = lines,
+        }
+        local buf_lines = vim.api.nvim_buf_get_lines(wbufnr, 0, -1, false)
+        for i, l in ipairs(buf_lines) do
+          if #vim.split(l, "pass") > 1 then
+            vim.api.nvim_buf_add_highlight(wbufnr, -1, "OctoPassingTest", i - 1, 0, -1)
+          elseif #vim.split(l, "fail") > 1 then
+            vim.api.nvim_buf_add_highlight(wbufnr, -1, "OctoFailingTest", i - 1, 0, -1)
+          end
+        end
+      end
+    end,
+  }
+end
+
+---@param params table
+---@param buffer OctoBuffer
+---@param bufnr integer MR IID
+---@param default_merge_method string
+function M.cmds_merge_pr(params, buffer, bufnr, default_merge_method)
+  local args = { "pr", "merge", tostring(buffer.number) }
+  for i = 1, params.n do
+    if params[i] == "delete" then
+      table.insert(args, "--delete-branch")
+    end
+  end
+  local has_flag = false
+  for i = 1, params.n do
+    if params[i] == "commit" then
+      table.insert(args, "--merge")
+      has_flag = true
+    elseif params[i] == "squash" then
+      table.insert(args, "--squash")
+      has_flag = true
+    elseif params[i] == "rebase" then
+      table.insert(args, "--rebase")
+      has_flag = true
+    end
+  end
+  if not has_flag then
+    if default_merge_method == "squash" then
+      table.insert(args, "--squash")
+    elseif default_merge_method == "rebase" then
+      table.insert(args, "--rebase")
+    else
+      table.insert(args, "--merge")
+    end
+  end
+
+  cli.run {
+    args = args,
+    cb = function(output, stderr)
+      utils.info(output .. " " .. stderr)
+      writers.write_state(bufnr)
+    end,
+  }
+end
+
+---@param buffer OctoBuffer
+function M.cmds_show_pr_diff(buffer)
+  local url = string.format("/repos/%s/pulls/%s", buffer.repo, buffer.number)
+  cli.run {
+    args = { "api", "--paginate", url },
+    headers = { "Accept: application/vnd.github.v3.diff" },
+    cb = function(output, stderr)
+      if stderr and not utils.is_blank(stderr) then
+        utils.error(stderr)
+      elseif output then
+        local lines = vim.split(output, "\n")
+        local wbufnr = vim.api.nvim_create_buf(true, true)
+        vim.api.nvim_buf_set_lines(wbufnr, 0, -1, false, lines)
+        vim.api.nvim_set_current_buf(wbufnr)
+        vim.api.nvim_buf_set_option(wbufnr, "filetype", "diff")
+      end
+    end,
+  }
+end
+
+---add/delete reaction
+---@param reaction string THUMBS_UP or THUMBS_DOWN or HOORAY
+---@param id string
+---@param buffer OctoBuffer
+---@param bufnr integer
+---@param action string add or remove
+---@param reaction_line integer
+---@param insert_line boolean
+function M.cmds_reaction_action(reaction, id, buffer, bufnr, action, reaction_line, insert_line)
+  local query = graphql(action .. "_reaction_mutation", id, reaction)
+
+  cli.run {
+    args = { "api", "graphql", "-f", string.format("query=%s", query) },
+    cb = function(output, stderr)
+      if stderr and not utils.is_blank(stderr) then
+        utils.error(stderr)
+      elseif output then
+        local resp = vim.fn.json_decode(output)
+        local reaction_groups
+        if action == "add" then
+          reaction_groups = resp.data.addReaction.subject.reactionGroups
+        elseif action == "remove" then
+          reaction_groups = resp.data.removeReaction.subject.reactionGroups
+        end
+
+        buffer:update_reactions_at_cursor(reaction_groups, reaction_line)
+        if action == "remove" and utils.count_reactions(reaction_groups) == 0 then
+          -- delete lines
+          vim.api.nvim_buf_set_lines(bufnr, reaction_line - 1, reaction_line + 1, false, {})
+          vim.api.nvim_buf_clear_namespace(bufnr, constants.OCTO_REACTIONS_VT_NS, reaction_line - 1, reaction_line + 1)
+        elseif action == "add" and insert_line then
+          -- add lines
+          vim.api.nvim_buf_set_lines(bufnr, reaction_line - 1, reaction_line - 1, false, { "", "" })
+        end
+        writers.write_reactions(bufnr, reaction_groups, reaction_line)
+        buffer:update_metadata()
+      end
+    end,
+  }
+end
+
+---@param buffer OctoBuffer
+---@param bufnr integer
+---@param column_id string
+function M.cmds_add_project_card(buffer, bufnr, column_id)
+  local query = graphql("add_project_card_mutation", buffer.node.id, column_id)
+
+  cli.run {
+    args = { "api", "graphql", "--paginate", "-f", string.format("query=%s", query) },
+    cb = function(output, stderr)
+      if stderr and not utils.is_blank(stderr) then
+        utils.error(stderr)
+      elseif output then
+        -- refresh issue/pr details
+        require("octo").load(buffer.repo, buffer.kind, buffer.number, function(obj)
+          writers.write_details(bufnr, obj, true)
+          buffer.node.projectCards = obj.projectCards
+        end)
+      end
+    end,
+  }
+end
+
+---@param buffer OctoBuffer
+---@param bufnr integer
+---@param card string
+function M.cmds_remove_project_card(buffer, bufnr, card)
+  local query = graphql("delete_project_card_mutation", card)
+
+  cli.run {
+    args = { "api", "graphql", "--paginate", "-f", string.format("query=%s", query) },
+    cb = function(output, stderr)
+      if stderr and not utils.is_blank(stderr) then
+        utils.error(stderr)
+      elseif output then
+        -- refresh issue/pr details
+        require("octo").load(buffer.repo, buffer.kind, buffer.number, function(obj)
+          buffer.node.projectCards = obj.projectCards
+          writers.write_details(bufnr, obj, true)
+        end)
+      end
+    end,
+  }
+end
+
+---@param buffer OctoBuffer
+---@param bufnr integer
+---@param source_card string
+---@param target_column string
+function M.cmds_move_project_card(buffer, bufnr, source_card, target_column)
+  local query = graphql("move_project_card_mutation", source_card, target_column)
+
+  cli.run {
+    args = { "api", "graphql", "--paginate", "-f", string.format("query=%s", query) },
+    cb = function(output, stderr)
+      if stderr and not utils.is_blank(stderr) then
+        utils.error(stderr)
+      elseif output then
+        -- refresh issue/pr details
+        require("octo").load(buffer.repo, buffer.kind, buffer.number, function(obj)
+          buffer.node.projectCards = obj.projectCards
+          writers.write_details(bufnr, obj, true)
+        end)
+      end
+    end,
+  }
+end
+
+---@param buffer OctoBuffer
+---@param bufnr integer
+---@param project_id string
+---@param field_id string
+---@param value string
+function M.cmds_set_project_card_v2(buffer, bufnr, project_id, field_id, value)
+  local add_query = graphql("add_project_v2_item_mutation", buffer.node.id, project_id)
+
+  cli.run {
+    args = { "api", "graphql", "--paginate", "-f", string.format("query=%s", add_query) },
+    cb = function(add_output, add_stderr)
+      if add_stderr and not utils.is_blank(add_stderr) then
+        utils.error(add_stderr)
+      elseif add_output then
+        local resp = vim.fn.json_decode(add_output)
+        local update_query = graphql(
+          "update_project_v2_item_mutation",
+          project_id,
+          resp.data.addProjectV2ItemById.item.id,
+          field_id,
+          value
+        )
+
+        cli.run {
+          args = { "api", "graphql", "--paginate", "-f", string.format("query=%s", update_query) },
+          cb = function(update_output, update_stderr)
+            if update_stderr and not utils.is_blank(update_stderr) then
+              utils.error(update_stderr)
+            elseif update_output then
+              -- TODO do update here
+              -- refresh issue/pr details
+              require("octo").load(buffer.repo, buffer.kind, buffer.number, function(obj)
+                writers.write_details(bufnr, obj, true)
+                buffer.node.projectCards = obj.projectCards
+              end)
+            end
+          end,
+        }
+      end
+    end,
+  }
+end
+
+---@param buffer OctoBuffer
+---@param bufnr integer
+---@param project_id string
+---@param item_id string
+function M.cmds_remove_project_card_v2(buffer, bufnr, project_id, item_id)
+  local query = graphql("delete_project_v2_item_mutation", project_id, item_id)
+
+  cli.run {
+    args = { "api", "graphql", "--paginate", "-f", string.format("query=%s", query) },
+    cb = function(output, stderr)
+      if stderr and not utils.is_blank(stderr) then
+        utils.error(stderr)
+      elseif output then
+        -- refresh issue/pr details
+        require("octo").load(buffer.repo, buffer.kind, buffer.number, function(obj)
+          buffer.node.projectCards = obj.projectCards
+          writers.write_details(bufnr, obj, true)
+        end)
+      end
+    end,
+  }
+end
+
+---@param buffer OctoBuffer
+---@param bufnr integer
+---@param name string
+---@param description string
+---@param color string
+function M.cmds_create_label(buffer, bufnr, name, description, color)
+  local repo_id = utils.get_repo_id(buffer.repo)
+  local query = graphql("create_label_mutation", repo_id, name, description, color)
+  cli.run {
+    args = { "api", "graphql", "-f", string.format("query=%s", query) },
+    cb = function(output, stderr)
+      if stderr and not utils.is_blank(stderr) then
+        utils.error(stderr)
+      elseif output then
+        local resp = vim.fn.json_decode(output)
+        local label = resp.data.createLabel.label
+        utils.info("Created label: " .. label.name)
+
+        -- refresh issue/pr details
+        require("octo").load(buffer.repo, buffer.kind, buffer.number, function(obj)
+          writers.write_details(bufnr, obj, true)
+        end)
+      end
+    end,
+  }
+end
+
+---@param buffer OctoBuffer
+---@param bufnr integer
+---@param iid string
+---@param label_id string
+function M.cmds_add_label(buffer, bufnr, iid, label_id)
+  local query = graphql("add_labels_mutation", iid, label_id)
+
+  cli.run {
+    args = { "api", "graphql", "-f", string.format("query=%s", query) },
+    cb = function(output, stderr)
+      if stderr and not utils.is_blank(stderr) then
+        utils.error(stderr)
+      elseif output then
+        -- refresh issue/pr details
+        require("octo").load(buffer.repo, buffer.kind, buffer.number, function(obj)
+          writers.write_details(bufnr, obj, true)
+        end)
+      end
+    end,
+  }
+end
+
+---@param buffer OctoBuffer
+---@param bufnr integer
+---@param iid string
+---@param label_id string
+function M.cmds_remove_label(buffer, bufnr, iid, label_id)
+  local query = graphql("remove_labels_mutation", iid, label_id)
+
+  cli.run {
+    args = { "api", "graphql", "-f", string.format("query=%s", query) },
+    cb = function(output, stderr)
+      if stderr and not utils.is_blank(stderr) then
+        utils.error(stderr)
+      elseif output then
+        -- refresh issue/pr details
+        require("octo").load(buffer.repo, buffer.kind, buffer.number, function(obj)
+          writers.write_details(bufnr, obj, true)
+        end)
+      end
+    end,
+  }
+end
+
+---assignee or reviewer dynamicly depending on subject
+---@param buffer OctoBuffer
+---@param bufnr integer
+---@param iid string
+---@param user_id string
+---@param subject string
+function M.cmds_add_user(buffer, bufnr, iid, user_id, subject)
+  local query
+  if subject == "assignee" then
+    query = graphql("add_assignees_mutation", iid, user_id)
+  elseif subject == "reviewer" then
+    query = graphql("request_reviews_mutation", iid, user_id)
+  else
+    utils.error "Invalid user type"
+    return
+  end
+
+  cli.run {
+    args = { "api", "graphql", "--paginate", "-f", string.format("query=%s", query) },
+    cb = function(output, stderr)
+      if stderr and not utils.is_blank(stderr) then
+        utils.error(stderr)
+      elseif output then
+        -- refresh issue/pr details
+        require("octo").load(buffer.repo, buffer.kind, buffer.number, function(obj)
+          writers.write_details(bufnr, obj, true)
+          vim.cmd [[stopinsert]]
+        end)
+      end
+    end,
+  }
+end
+
+---@param buffer OctoBuffer
+---@param bufnr integer
+---@param iid string
+---@param user_id string
+function M.cmds_remove_assignee(buffer, bufnr, iid, user_id)
+  local query = graphql("remove_assignees_mutation", iid, user_id)
+
+  cli.run {
+    args = { "api", "graphql", "--paginate", "-f", string.format("query=%s", query) },
+    cb = function(output, stderr)
+      if stderr and not utils.is_blank(stderr) then
+        utils.error(stderr)
+      elseif output then
+        -- refresh issue/pr details
+        require("octo").load(buffer.repo, buffer.kind, buffer.number, function(obj)
+          writers.write_details(bufnr, obj, true)
+        end)
+      end
+    end,
+  }
+end
 
 return M
